requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-QRcode==2.0.2
Frozen-Flask==0.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lyft-rides==0.2
MarkupSafe==1.1.1
Pillow==6.0.0
PyYAML==5.1
qrcode==6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
